Move lib/api to modules/api
